JService now inherits from JComponent
