+++ src/libraries/JANA/JService.h
@@ -1,0 +1,5 @@
+
+#pragma once
+#include <JANA/JServiceFwd.h>
+#include <JANA/Omni/JComponent.h>
+
